refactor: move utils functions to separate folder
